feat: implement product creation and update functionality, enhance navigation and authentication checks

+++ ecommerceBackend9/src/pages/admin/dashboard/product/Product.jsx
@@ -1,44 +1,46 @@
-import React, { useEffect, useState } from 'react'
-import { handleDeleteOperation, handleGetOperation } from '../../../../handleOperation/handleOperation'
-import { apiLinks } from '../../../../handleOperation/apiLinks'
+import React, { useEffect, useState } from "react";
+import {
+  handleDeleteOperation,
+  handleGetOperation,
+} from "../../../../handleOperation/handleOperation";
+import { apiLinks } from "../../../../handleOperation/apiLinks";
+import { Link, useNavigate } from "react-router-dom";
 
 const Product = () => {
+  const [formData, setFormData] = useState([]);
+  const [count, setCount] = useState(1);
 
-  const [formData, setFormData] = useState([])
-
+  const navigate = useNavigate();
 
   useEffect(() => {
     const fetchProduct = async () => {
-
       try {
-
-        const result = await handleGetOperation(apiLinks.getAllProduct)
+        const result = await handleGetOperation(apiLinks.getAllProduct);
 
         if (result.status == 200) {
-
-
-          setFormData(result.data.data)
+          setFormData(result.data.data);
         }
       } catch (error) {
-        alert("Failed to fetch product")
+        alert("Failed to fetch product");
       }
-
-    }
-    fetchProduct()
-  }, [])
-
+    };
+    fetchProduct();
+  }, [count]);
 
   const handledDelete = async (id) => {
-    alert(id)
-    const result = await handleDeleteOperation(apiLinks.deleteProduct(id))
+    const isDelete = confirm("Do you want to delete ");
+    if (isDelete) {
+      const result = await handleDeleteOperation(apiLinks.deleteProduct(id));
 
-    console.log(result);
-
-  }
+      if (result.status == 200) {
+        setCount((prev) => prev + 1);
+        // location.href = '/admin/product'
+      }
+    }
+  };
 
   return (
     <div>
-
       <div className="flex flex-col overflow-y-auto h-full">
         <table className="min-w-full bg-white border undefined">
           <thead className="text-xs text-gray-700 uppercase bg-gray-50 h-6 sticky top-0 z-10 transition-all ease-in-out ">
@@ -52,47 +54,59 @@ const Product = () => {
             </tr>
           </thead>
           <tbody className="undefined select-none">
-
             {formData.map((product) => (
               <tr className="border-b" key={product._id}>
-
-
                 <td className="px-6 py-1.5 whitespace-nowrap text-sm">
                   <div className="flex items-center gap-2">
-                    <img src={product.imageUrl} alt="image" className="aspect-square rounded-lg object-cover h-16" />
+                    <img
+                      src={product.imageUrl}
+                      alt="image"
+                      className="aspect-square rounded-lg object-cover h-16"
+                    />
                     <p className="font-bold">{product.productName}</p>
                   </div>
                 </td>
-                <td className="px-6 py-1.5 whitespace-nowrap text-sm  text-center">{product.brand}</td>
+                <td className="px-6 py-1.5 whitespace-nowrap text-sm  text-center">
+                  {product.brand}
+                </td>
                 <td className="px-6 py-1.5 whitespace-nowrap text-sm   flex justify-center items-center h-full gap-2">
                   <td className="px-6 py-1.5 whitespace-nowrap text-sm  text-center">
-                    <button className="text-xs font-bold me-2 px-2.5 shadow py-2  bg-green-200 text-green-600 font-semibold rounded-lg shadow">edit</button>
-                    <button onClick={() => (handledDelete(product._id))} className="text-xs font-bold me-2 px-2.5 shadow py-2  bg-red-500 text-white font-semibold rounded-lg shadow">delete</button>
+                    <Link
+                      to={`/admin/updateProduct/${product._id}`}
+                      className="text-xs font-bold me-2 px-2.5 shadow py-2  bg-green-200 text-green-600 font-semibold rounded-lg shadow"
+                    >
+                      edit
+                    </Link>
+                    <button
+                      onClick={() => handledDelete(product._id)}
+                      className="text-xs font-bold me-2 px-2.5 shadow py-2  bg-red-500 text-white font-semibold rounded-lg shadow"
+                    >
+                      delete
+                    </button>
                   </td>
                   <div className="relative inline-block text-left select-none ">
-                    <div className="rounded-full py-4 hover:bg-gray-100 focus:outline- cursor-pointer "><svg xmlns="http://www.w3.org/2000/svg" width={26} height={26} viewBox="0 0 24 24">
-                      <path fill="currentColor" d="M12 16a2 2 0 0 1 2 2a2 2 0 0 1-2 2a2 2 0 0 1-2-2a2 2 0 0 1 2-2m0-6a2 2 0 0 1 2 2a2 2 0 0 1-2 2a2 2 0 0 1-2-2a2 2 0 0 1 2-2m0-6a2 2 0 0 1 2 2a2 2 0 0 1-2 2a2 2 0 0 1-2-2a2 2 0 0 1 2-2">
-                      </path>
-                    </svg></div>
+                    <div className="rounded-full py-4 hover:bg-gray-100 focus:outline- cursor-pointer ">
+                      <svg
+                        xmlns="http://www.w3.org/2000/svg"
+                        width={26}
+                        height={26}
+                        viewBox="0 0 24 24"
+                      >
+                        <path
+                          fill="currentColor"
+                          d="M12 16a2 2 0 0 1 2 2a2 2 0 0 1-2 2a2 2 0 0 1-2-2a2 2 0 0 1 2-2m0-6a2 2 0 0 1 2 2a2 2 0 0 1-2 2a2 2 0 0 1-2-2a2 2 0 0 1 2-2m0-6a2 2 0 0 1 2 2a2 2 0 0 1-2 2a2 2 0 0 1-2-2a2 2 0 0 1 2-2"
+                        ></path>
+                      </svg>
+                    </div>
                   </div>
                 </td>
               </tr>
             ))}
-
-
-
-
-
-
-
-
           </tbody>
         </table>
       </div>
-
-
     </div>
-  )
-}
+  );
+};
 
-export default Product
+export default Product;
